refactor: 클래스 위치 변경에 따른 위치 변경
