+++ src/main/java/com/twentyone/steachserver/domain/auth/service/AuthServiceImpl.java
@@ -9,7 +9,7 @@ import com.twentyone.steachserver.domain.member.model.Student;
 import com.twentyone.steachserver.domain.member.model.Teacher;
 import com.twentyone.steachserver.domain.member.repository.StudentRepository;
 import com.twentyone.steachserver.domain.member.repository.TeacherRepository;
-import com.twentyone.steachserver.util.FileUtil;
+import com.twentyone.steachserver.util.converter.FileUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
