+++ src/main/java/com/twentyone/steachserver/domain/studentLecture/model/StudentLecture.java
@@ -1,6 +1,6 @@
 package com.twentyone.steachserver.domain.studentLecture.model;
 
-import com.twentyone.steachserver.config.domain.BaseTimeEntity;
+import com.twentyone.steachserver.domain.common.BaseTimeEntity;
 import com.twentyone.steachserver.domain.lecture.model.Lecture;
 import com.twentyone.steachserver.domain.member.model.Student;
 import jakarta.persistence.*;
